+++ components/ui/Sparkles.tsx
@@ -40,7 +40,6 @@ const Sparkles = (props: ParticlesProps) => {
 
   const particlesLoaded = async (container?: Container) => {
     if (container) {
-      console.log(container);
       controls.start({
         opacity: 1,
         transition: {